File rename nrrd/axes.c --> nrrd/axis.c, as part of last great rename...
This fixes a long-standing inconsistency in the names of functions,
enum values, an airEnum, and a struct, which are involved in
representation of axis meta-information.

The happy consequence of this renamaing is that "nrrdAxes..." now ONLY
refers to operations on the in-memory ordering of samples in the
volume (e.g., nrrdAxesPermute, nrrdAxesSplit), even if superficially,
only one axis is affected (e.g., nrrdAxesDelete).  Actually, from the
standpoint of in-memory ordering, any operation on one axis affects
either all axes or no axes, so there is a logic here.  These functions
have actually kept their names.

"nrrdAxisInfo...", on the other hand, refers to the representation and
manipulation of the meta-information that nrrd maintains for each
axis.  This is where the renaming happened:

          airEnum *nrrdAxesInfo --> nrrdAxisInfo
    nrrdAxesInfo... enum values --> nrrdAxisInfo...
      #define NRRD_AXESINFO_MAX --> NRRD_AXIS_INFO_MAX
  #define NRRD_AXESINFO_..._BIT --> NRRD_AXIS_INFO_..._BIT
   nrrdAxes{Copy,Set...,Get...} --> nrrdAxisInfo{Copy,Set...,Get...}

Note that the struct NrrdAxis has kept its name, as opposed to being
renamed struct NrrdAxisInfo, because:
1) I didn't want to mess with anything inside the main Nrrd struct
   (there is an array of NrrdAxis structs in the Nrrd).
2) I wanted to avoid pointless confusion between airEnum *nrrdAxisInfo
   and a C struct NrrdAxisInfo.
3) 99.9% of Nrrd users never see or type "NrrdAxis" in their code-- but
   the nrrdAxisInfo... function names and the airEnum name are quite
   common.

Thus, "axisinfo" (in some form of capitalization) is now used for what
formerly was either "axes", "axesinfo", or "axis".  This element of
confusion bit me for the last time this morning.