This is a functionality change relating to use of pre-allocated data in nrrdRead().  There is no public API change.
Prior to this change, a nrrd that was already already valid and allocated
with some amount of memory would be stripped of that memory (the memory
was freed) by nrrdRead().  The various encodings would allocate new memory
via _nrrdCalloc.  This meant that even when the caller to nrrdRead knew
in advance that the given nrrd could hold the data that would be read from
the given file, the memory was freed and re-allocated.  This presented a
problem for ITK's use of nrrd in NrrdIO.

Now, near the beginning of nrrdRead(), any existing memory (via nrrd->data
being non-NULL) and its allocated size (via nrrd->type and nrrd->axis[i].size)
are remembered in the NrrdIoState struct (void *oldData, size_t oldDataSize),
which is then referred to via _nrrdCalloc, or directly by the format's reader.
The result is that existing memory is used whenever the allocated size
matches the size described by the new nrrd's header.  This change has been
made in all existing formats, which involved more modifications for
formatText.c and formatPNG.c, since these do not use (nor should they)
functions in the existing encodingXXX.c files.

Near the end of nrrdRead(), the old data is freed if it turns out that it
wasn't re-used.