API CHANGE, API RENAME:
some functions have been renamed and their arguments are different:
nrrdPeripheralInit, nrrdPeripheralCopy -> nrrdBasicInfoInit, nrrdBasicInfoCopy

The new functions take an "exclusion bitflag", exactly like
nrrdAxisInfoCopy.  The "basic info" bitflags are based on the values
in the *new* nrrdBasicInfo airEnum.  Using the bitflags allows much
finer-grained control over which values are initialized or copied at
each operation.  I've followed through with these changes in all the
Nrrd code that goes into NrrdIO (used in ITK), but have yet to finish
carefully propagating it all to the rest of nrrd.

Actually, for the time being, nrrdPeripheralInit and nrrdPeripheralCopy
are still there, but now they're declared in privateNrrd.h, not in
nrrd.h.